Refactor/Redesign Method Tab
Removed backpack, canoe/barge, & raft/boats widgets and associated
tables in database.

Added new “Equipment” table to database with fields from the three
tables that were removed as well as the method-related fields in the
“SamplingEvents” feature class. Removed those fields from the
“SamplingEvents” feature class.

Added new field “TYPE” and associated domain to the “Equipment” table.

Added equipment widget to the Method tab to match the new equipment
table in the database.

Closes #2