Database Changes (#133)
* Tags table schema changes

These schema changes are outlined in #72

I also removed a few duplicate domains that I found.

The *_1 domains were the updated versions. I deleted the old domains and renamed the new ones to remove the _1.

* Change MEASUREMENT field type from SHORT to DOUBLE

Related to #116

* overwrite all domain codes with corresponding descriptions

Adjust field lengths to match the length of the longest description. Translate any data in the tables as well.

I didn't have to worry about excluding the species code or length type domains since the codes already matched the descriptions.

Closes #97