stop isTrusted from being added to row
addRow was being called with a click event which passed an event to it's default parameter.

This was unexpected since you probably only thought data should be sent to the function.

I added another parameter to capture and ignore the click event and passed the data as the second optional parameter.